worker base classes cleanup
